+++ packages/server-core/src/projects/portal/portal.hooks.ts
@@ -22,7 +22,7 @@ import authenticate from '../../hooks/authenticate'
 
 export default {
   around: {
-    all: []
+    all: [authenticate()]
   },
 
   before: {
