+++ packages/server-core/src/projects/scene-data/scene-data.hooks.ts
@@ -22,7 +22,7 @@ import authenticate from '../../hooks/authenticate'
 
 export default {
   around: {
-    all: []
+    all: [authenticate()]
   },
 
   before: {
