+++ packages/server-core/src/projects/scene-upload/scene-upload.hooks.ts
@@ -23,7 +23,7 @@ import verifyScope from '../../hooks/verify-scope'
 
 export default {
   around: {
-    all: []
+    all: [authenticate()]
   },
 
   before: {
